+++ cpp/tests/replace/replace_nulls_tests.cu
@@ -38,7 +38,7 @@ TEST_F(ReplaceErrorTest, SizeMismatch)
                                                                {0, 0, 1, 1, 1, 1, 1, 1}};
   cudf::test::fixed_width_column_wrapper<int32_t> values_to_replace_column{{10, 11, 12, 13}};
 
-  ASSERT_THROW(cudf::experimental::replace_nulls(input_column, values_to_replace_column, mr()),
+  ASSERT_THROW(cudf::replace_nulls(input_column, values_to_replace_column, mr()),
                cudf::logic_error);
 }
 
@@ -50,7 +50,7 @@ TEST_F(ReplaceErrorTest, TypeMismatch)
   cudf::test::fixed_width_column_wrapper<float> values_to_replace_column{
     {10, 11, 12, 13, 14, 15, 16, 17}};
 
-  EXPECT_THROW(cudf::experimental::replace_nulls(input_column, values_to_replace_column, mr()),
+  EXPECT_THROW(cudf::replace_nulls(input_column, values_to_replace_column, mr()),
                cudf::logic_error);
 }
 
@@ -61,8 +61,7 @@ TEST_F(ReplaceErrorTest, TypeMismatchScalar)
                                                                {0, 0, 1, 1, 1, 1, 1, 1}};
   cudf::numeric_scalar<float> replacement(1);
 
-  EXPECT_THROW(cudf::experimental::replace_nulls(input_column, replacement, mr()),
-               cudf::logic_error);
+  EXPECT_THROW(cudf::replace_nulls(input_column, replacement, mr()), cudf::logic_error);
 }
 
 struct ReplaceNullsStringsTest : public cudf::test::BaseFixture {
@@ -82,7 +81,7 @@ TEST_F(ReplaceNullsStringsTest, SimpleReplace)
     replacement.begin(), replacement.end(), replacement_v.begin()};
 
   std::unique_ptr<cudf::column> result;
-  ASSERT_NO_THROW(result = cudf::experimental::replace_nulls(input_w, replacement_w, mr()));
+  ASSERT_NO_THROW(result = cudf::replace_nulls(input_w, replacement_w, mr()));
 
   cudf::test::expect_columns_equal(*result, expected_w);
 }
@@ -101,7 +100,7 @@ TEST_F(ReplaceNullsStringsTest, ReplaceWithNulls)
     replacement.begin(), replacement.end(), replacement_v.begin()};
 
   std::unique_ptr<cudf::column> result;
-  ASSERT_NO_THROW(result = cudf::experimental::replace_nulls(input_w, replacement_w, mr()));
+  ASSERT_NO_THROW(result = cudf::replace_nulls(input_w, replacement_w, mr()));
 
   cudf::test::expect_columns_equal(*result, expected_w);
 }
@@ -119,7 +118,7 @@ TEST_F(ReplaceNullsStringsTest, ReplaceWithAllNulls)
   cudf::test::strings_column_wrapper expected_w{input.begin(), input.end(), input_v.begin()};
 
   std::unique_ptr<cudf::column> result;
-  ASSERT_NO_THROW(result = cudf::experimental::replace_nulls(input_w, replacement_w, mr()));
+  ASSERT_NO_THROW(result = cudf::replace_nulls(input_w, replacement_w, mr()));
 
   cudf::test::expect_columns_equal(*result, expected_w);
 }
@@ -137,7 +136,7 @@ TEST_F(ReplaceNullsStringsTest, ReplaceWithAllEmpty)
   cudf::test::strings_column_wrapper expected_w{input.begin(), input.end(), replacement_v.begin()};
 
   std::unique_ptr<cudf::column> result;
-  ASSERT_NO_THROW(result = cudf::experimental::replace_nulls(input_w, replacement_w, mr()));
+  ASSERT_NO_THROW(result = cudf::replace_nulls(input_w, replacement_w, mr()));
 
   cudf::test::expect_columns_equal(*result, expected_w);
 }
@@ -155,7 +154,7 @@ TEST_F(ReplaceNullsStringsTest, ReplaceNone)
   cudf::test::strings_column_wrapper expected_w{input.begin(), input.end()};
 
   std::unique_ptr<cudf::column> result;
-  ASSERT_NO_THROW(result = cudf::experimental::replace_nulls(input_w, replacement_w, mr()));
+  ASSERT_NO_THROW(result = cudf::replace_nulls(input_w, replacement_w, mr()));
 
   cudf::test::expect_columns_equal(*result, expected_w);
 }
@@ -172,7 +171,7 @@ TEST_F(ReplaceNullsStringsTest, SimpleReplaceScalar)
   cudf::test::strings_column_wrapper expected_w{expected.begin(), expected.end()};
 
   std::unique_ptr<cudf::column> result;
-  ASSERT_NO_THROW(result = cudf::experimental::replace_nulls(input_w, *repl, mr()));
+  ASSERT_NO_THROW(result = cudf::replace_nulls(input_w, *repl, mr()));
 
   cudf::test::expect_columns_equal(*result, expected_w);
 }
@@ -191,7 +190,7 @@ void ReplaceNullsColumn(cudf::test::fixed_width_column_wrapper<T> input,
                         cudf::test::fixed_width_column_wrapper<T> expected)
 {
   std::unique_ptr<cudf::column> result;
-  ASSERT_NO_THROW(result = cudf::experimental::replace_nulls(input, replacement_values));
+  ASSERT_NO_THROW(result = cudf::replace_nulls(input, replacement_values));
   expect_columns_equal(expected, *result);
 }
 
@@ -201,7 +200,7 @@ void ReplaceNullsScalar(cudf::test::fixed_width_column_wrapper<T> input,
                         cudf::test::fixed_width_column_wrapper<T> expected)
 {
   std::unique_ptr<cudf::column> result;
-  ASSERT_NO_THROW(result = cudf::experimental::replace_nulls(input, replacement_value));
+  ASSERT_NO_THROW(result = cudf::replace_nulls(input, replacement_value));
   expect_columns_equal(expected, *result);
 }